feat(dashboard): ✨ Enhance Flight Crew Member Dashboard with incident severity metrics
- Refactored queries to count legs by severity levels into separate methods for low, medium, and high incidents.
- Updated `FlightCrewMemberDashboardShowService` to utilize new methods for fetching incident counts.
- Adjusted data types for deviation metrics in `FlightCrewMemberDashboard` to use `Double`.
- Added new labels in English and Spanish for the dashboard to reflect the changes in incident metrics.

+++ src/main/java/acme/forms/flightCrewMemberDashboard/FlightCrewMemberDashboard.java
@@ -25,9 +25,9 @@ public class FlightCrewMemberDashboard extends AbstractForm {
 	private List<String>		confirmedFlightAssignments;
 	private List<String>		pendingFlightAssignments;
 	private List<String>		cancelledFlightAssignments;
-	private Integer				averageDeviationNumberOfFlights;
+	private Double				averageDeviationNumberOfFlights;
 	private Integer				minimumDeviationNumberOfFlights;
 	private Integer				maximumDeviationNumberOfFlights;
-	private Integer				standardDeviationNumberOfFlights;
+	private Double				standardDeviationNumberOfFlights;
 
 }
